requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
Pillow==10.3.0
aiofiles==23.2.1
# Optional: for AI_SERVICE=openclip (local vision model)
torch>=2.0.0
open-clip-torch>=2.24.0
# Object detection (YOLOv5s for main-object cropping)
ultralytics>=8.1.0
# Classifier head on embeddings
scikit-learn>=1.4.0
numpy>=1.24.0
# Listing text embeddings (TEXT_EMBEDDING_PROVIDER=sentence_transformers); not used for image/OpenCLIP
sentence-transformers>=2.2.0
python-dotenv==1.0.1
httpx==0.27.0
pytest>=7.0.0
